自动嵌入或者还原
