+++ newDesktop/newDesktop/main.h
@@ -1,1 +1,2 @@
 #pragma once
+#include"putInside.h"
