+++ newDesktop/newDesktop/newDesktop.vcxproj
@@ -116,9 +116,11 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="main.h" />
+    <ClInclude Include="putInside.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="mian.cpp" />
+    <ClCompile Include="putInside.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
