+++ newDesktop/newDesktop/newDesktop.vcxproj.filters
@@ -18,9 +18,15 @@
     <ClInclude Include="main.h">
       <Filter>头文件</Filter>
     </ClInclude>
+    <ClInclude Include="putInside.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="mian.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="putInside.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
   </ItemGroup>
